pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "This is a small minimal AI inference API for image classification"
authors = ["Anas B <anasbachiri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic = "^2.0.0"
requests = "^2.32.0"
loguru = "^0.7.0"
joblib = "^1.2.0"
python-multipart = "^0.0.20"
scikit-learn = "^1.5.0"
pillow = "^10.2.0"
torch = "2.5.1"
torchaudio = "2.5.1"
torchvision = "0.20.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = {version = "^24.3", allow-prereleases = true}
autopep8 = "^2.0.0"
ipdb = "^0.13.0"
pylint = "^2.15.0"  # Fixed to a valid version

[tool.poetry.group.aws]
optional = true

[tool.poetry.group.aws.dependencies]
mangum = "^0.14.0"

#[[tool.poetry.source]]
#name = "torch"
#url = "https://download.pytorch.org/whl/cu118"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''
